feat: remove unused part, use reference for parameter if possible, and remove virtual from child class
