+++ include/aruku/config/grpc/call_data_save_config.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2023 Ichiro ITS
+// Copyright (c) 2024 Ichiro ITS
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -32,10 +32,10 @@ class CallDataSaveConfig
 public:
   CallDataSaveConfig(
     aruku_interfaces::proto::Config::AsyncService * service, grpc::ServerCompletionQueue * cq,
-    const std::string path);
+    const std::string & path);
 
 protected:
-  virtual void AddNextToCompletionQueue() override;
+  void AddNextToCompletionQueue() override;
   void WaitForRequest();
   void HandleRequest();
 };
