+++ src/aruku/config/grpc/call_data_get_config.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2023 Ichiro ITS
+// Copyright (c) 2024 Ichiro ITS
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -20,16 +20,16 @@
 
 #include "aruku/config/grpc/call_data_get_config.hpp"
 
+#include "aruku/config/utils/config.hpp"
 #include "aruku_interfaces/aruku.grpc.pb.h"
 #include "aruku_interfaces/aruku.pb.h"
-#include "aruku/config/utils/config.hpp"
 #include "rclcpp/rclcpp.hpp"
 
 namespace aruku
 {
 CallDataGetConfig::CallDataGetConfig(
   aruku_interfaces::proto::Config::AsyncService * service, grpc::ServerCompletionQueue * cq,
-  const std::string path)
+  const std::string & path)
 : CallData(service, cq, path)
 {
   Proceed();
@@ -45,7 +45,6 @@ void CallDataGetConfig::WaitForRequest()
 void CallDataGetConfig::HandleRequest()
 {
   Config config(path_);
-
   reply_.set_json_kinematic(config.get_config("kinematic"));
   reply_.set_json_walking(config.get_config("walking"));
   RCLCPP_INFO(rclcpp::get_logger("Get config"), "config has been sent!");
